# speed up CSS class queries