--- a/ext/java/nokogiri/internals/NokogiriXPathFunction.java
+++ b/ext/java/nokogiri/internals/NokogiriXPathFunction.java
@@ -36,6 +36,7 @@ import java.util.List;
 
 import javax.xml.xpath.XPathFunction;
 import javax.xml.xpath.XPathFunctionException;
+import javax.xml.namespace.QName;
 
 import org.jruby.Ruby;
 import org.jruby.RubyArray;
@@ -64,14 +65,14 @@ import static nokogiri.internals.NokogiriHelpers.nodeListToRubyArray;
 public class NokogiriXPathFunction implements XPathFunction {
 
     private final IRubyObject handler;
-    private final String name;
+    private final QName name;
     private final int arity;
     
-    public static NokogiriXPathFunction create(IRubyObject handler, String name, int arity) {
+    public static NokogiriXPathFunction create(IRubyObject handler, QName name, int arity) {
         return new NokogiriXPathFunction(handler, name, arity);
     }
 
-    private NokogiriXPathFunction(IRubyObject handler, String name, int arity) {
+    private NokogiriXPathFunction(IRubyObject handler, QName name, int arity) {
         this.handler = handler;
         this.name = name;
         this.arity = arity;
@@ -82,11 +83,20 @@ public class NokogiriXPathFunction implements XPathFunction {
             throw new XPathFunctionException("arity does not match");
         }
         
+        if (name.getNamespaceURI().equals(NokogiriNamespaceContext.NOKOGIRI_BUILTIN_URI)) {
+            if (name.getLocalPart().equals("css-class")) {
+                return builtinCssClass(args);
+            }
+        }
+
+        if (this.handler.isNil()) {
+            throw new XPathFunctionException("no custom function handler declared for '" + name + "'");
+        }
+
         final Ruby runtime = this.handler.getRuntime();
         ThreadContext context = runtime.getCurrentContext();
-
-        IRubyObject result = Helpers.invoke(context, this.handler, this.name, fromObjectToRubyArgs(runtime, args));
-
+        IRubyObject result = Helpers.invoke(context, this.handler, this.name.getLocalPart(),
+                                            fromObjectToRubyArgs(runtime, args));
         return fromRubyToObject(runtime, result);
     }
 
@@ -120,5 +130,51 @@ public class NokogiriXPathFunction implements XPathFunction {
             return XmlNodeSet.newNodeSet(runtime, ((RubyArray) obj).toJavaArray());
         }
         /*if (o instanceof XmlNode)*/ return ((XmlNode) obj).getNode();
+    }
+
+    private static boolean builtinCssClass(List args) throws XPathFunctionException {
+        if (args.size() != 2) {
+            throw new XPathFunctionException("builtin function nokogiri:css-class takes two arguments");
+        }
+
+        String hay = args.get(0).toString();
+        String needle = args.get(1).toString();
+
+        if (needle.length() == 0) {
+            return true;
+        }
+
+        int j = 0;
+        int j_lim = hay.length() - needle.length();
+        while (j <= j_lim) {
+            int k;
+            for (k = 0; k < needle.length(); k++) {
+                if (needle.charAt(k) != hay.charAt(j+k)) {
+                    break;
+                }
+            }
+            if (k == needle.length()) {
+                if ((hay.length() == (j+k)) || isWhitespace(hay.charAt(j+k))) {
+                    return true ;
+                }
+            }
+
+            /* advance str to whitespace */
+            while (j <= j_lim && !isWhitespace(hay.charAt(j))) {
+                j++;
+            }
+
+            /* advance str to start of next word or end of string */
+            while (j <= j_lim && isWhitespace(hay.charAt(j))) {
+                j++;
+            }
+        }
+
+        return false;
+    }
+
+    private static boolean isWhitespace(char subject) {
+        // see libxml2's xmlIsBlank_ch()
+        return ((subject == 0x09) || (subject == 0x0A) || (subject == 0x0D) || (subject == 0x20));
     }
 }